requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
socksio==1.0.0
pydantic>=2.12
python-dotenv==1.1.1
jinja2==3.1.6
python-multipart==0.0.20
itsdangerous==2.2.0
cryptography==45.0.6